Move trans_field_ptr and struct_field_ptr to mir/lvalue
